moved shared components into shared directories and gitignored properly
